requirements.txt
fastapi
pyjwt[crypto]
httpx
pyyaml
uvicorn
requests
python-dotenv